+++ module/actor/actor.js
@@ -1199,7 +1199,7 @@ export class ArM5ePCActor extends Actor {
 
   // Utility functions
 
-  // get the XP coefficient of a given ability if any
+  // get the XP bonus of a given ability if any
 
   _getAbilityXpBonus(abilityKey = "", option = "") {
     if (abilityKey === "" || CONFIG.ARM5E.ALL_ABILITIES[abilityKey] == undefined) {
@@ -1218,7 +1218,7 @@ export class ArM5ePCActor extends Actor {
 
     return this.system.bonuses.skills[abilityKey].xpMod || 0;
   }
-
+  // get the XP coefficient of a given ability if any
   _getAbilityXpCoeff(abilityKey = "", option = "") {
     if (abilityKey === "" || CONFIG.ARM5E.ALL_ABILITIES[abilityKey] == undefined) {
       return 1.0;